# Add v8-compile-cache to make faster startup