--- a/marp-cli.js
+++ b/marp-cli.js
@@ -2,6 +2,7 @@
 
 'use strict'
 
+require('v8-compile-cache')
 require('./lib/marp-cli.js')
   .default(process.argv.slice(2))
   .then(exitCode => process.on('exit', () => process.exit(exitCode)))