Make CLI be faster by adding `v8-compile-cache` that attaches the require hook for using V8 code cache.

Benchmark of `time ./marp-cli.js` in my environments shows not dramatic but steady x1.15 speed up. (Use average time of 10 executions)

|`real`|Mac|Win|
|:---:|:---|:---|
|`master`|0.671s|0.960s|
|`v8-compile-cache`|0.585s (x1.15)|0.843s (x1.14)|